Fix: Foram incluídas algumas validações na customer.

+++ MiniERP/Services/SupplierService.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using MiniERP.Data;
 using MiniERP.Repositories;
+using MiniERP.Tools;
 
 namespace MiniERP.Services
 {
@@ -24,18 +25,49 @@ namespace MiniERP.Services
             return await _repository.GetSupplierByIdAsync(id);
         }
 
-        public async Task AddSupplierAsync(Supplier supplier)
+        public async Task AddSupplierAsync(ReqSupplier reqSupplier)
         {
+            var _sanitize = new Sanitize();
+
+            if (string.IsNullOrWhiteSpace(reqSupplier.Name))
+            {
+                throw new Exception("Você deve entrar com um nome válido!");
+            }
+
+            reqSupplier.Name = reqSupplier.Name.Trim();
+
+            var sanitizedName = _sanitize.ClearString(reqSupplier.Name);
+
+            if (sanitizedName != reqSupplier.Name)
+            {
+                throw new Exception("O nome do fornecedor contém caracteres inválidos. Apenas letras são permitidas.");
+            }
+
+            Supplier supplier = new Supplier();
+            supplier.Name= sanitizedName;
+
             await _repository.AddSupplierAsync(supplier);
         }
 
         public async Task UpdateSupplierAsync(Supplier supplier)
         {
+            Supplier respSupplier = await GetSupplierByIdAsync(supplier.SupplierId);
+            if (respSupplier != null)
+            {
+                throw new Exception("Fornecedor não encontrado!");
+            }
+
             await _repository.UpdateSupplierAsync(supplier);
         }
 
         public async Task DeleteSupplierAsync(int id)
         {
+            Supplier respSupplier = await GetSupplierByIdAsync(id);
+            if (respSupplier != null)
+            {
+                throw new Exception("Fornecedor não encontrado!");
+            }
+
             await _repository.DeleteSupplierAsync(id);
         }
     }
